Valid range is float for monthly std var

+++ seaice_ecdr/monthly.py
@@ -366,7 +366,7 @@ def calc_stdv_of_cdr_seaice_conc_monthly(
 
     stdv_of_cdr_seaice_conc_monthly = stdv_of_cdr_seaice_conc_monthly.assign_attrs(
         long_name="Passive Microwave Monthly Northern Hemisphere Sea Ice Concentration Source Estimated Standard Deviation",
-        valid_range=(0.0, 1.0),
+        valid_range=(np.float32(0.0), np.float32(1.0)),
         grid_mapping="crs",
     )
